[AP]: merge matmul_variadic_tpl
