--- python/paddle/apy/matmul_pass/matmul/composable_kernel
@@ -1,1 +1,0 @@
-/work/ck-5.7.1
