--- python/paddle/apy/matmul_pass/matmul/cutlass-3.7.0
@@ -1,1 +1,0 @@
-/paddle-workspace/cutlass-3.7.0
